dom/network/tests/test_network_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Network API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Network API **/

ok('mozConnection' in navigator, "navigator.mozConnection should exist");

</script>
</pre>
</body>
</html>
